requirements.txt
# Azure SDK packages
azure-cosmos>=4.5.0
azure-identity>=1.15.0
azure-mgmt-cosmosdb>=9.0.0

# OpenAI SDK
openai>=1.0.0

# Token counting
tiktoken>=0.5.0
